changing the return type of makeOrder endpoint to return clientSecret

+++ src/main/java/com/ihab/e_commerce/rest/request/stripe_payment/StripePaymentRequest.java
@@ -3,7 +3,7 @@ package com.ihab.e_commerce.rest.request.stripe_payment;
 import java.math.BigDecimal;
 
 public record StripePaymentRequest(
-    Long amount,
+    BigDecimal amount,
     String currency,
     String description
 ) {
